Add Form Validation to Send Form
Split WalletSendPage into a Component and Page
